seperate spinconfig into a seperate file, fixed import
