+++ monte_carlo/__init__.py
@@ -2,6 +2,7 @@
 
 # Add imports here
 from .functions import *
+from .spinconfig import SpinConfig
 
 # Handle versioneer
 from ._version import get_versions
